+++ demo/ARContainerViewController.swift
@@ -11,9 +11,8 @@ final class ARContainerViewController: UIViewController {
     private lazy var mapViewController = ARMapNavigationController()
     private lazy var arViewController = VisionARViewController()
 
-    weak var navigationDelegate: NavigationManagerDelegate?
+    weak var navigationDelegate: NavigationDelegate?
     private var navigationService: NavigationService?
-    private var routeHasChanged: Bool = true
     
     override func viewDidLoad() {
         mapViewController.completion = present
@@ -46,7 +45,9 @@ final class ARContainerViewController: UIViewController {
 
         navigationService = MapboxNavigationService(route: route)
         navigationService?.delegate = self
+        navigationDelegate?.navigation(didUpdate: Route(route: route))
         present(viewController: arViewController)
+        navigationService?.start()
     }
     
     func present(sampleBuffer: CMSampleBuffer) {
@@ -93,25 +94,21 @@ extension ARContainerViewController: NavigationServiceDelegate {
     func navigationService(_ service: NavigationService, didUpdate progress: RouteProgress, with location: CLLocation, rawLocation: CLLocation) {
         instructionsLabel.isHidden = false
         instructionsLabel.text = progress.currentLegProgress.currentStep.instructions
+    }
 
-        if routeHasChanged {
-            routeHasChanged = false
-            navigationDelegate?.navigationService(didUpdate: Route(route: progress.route))
+    func navigationService(_ service: NavigationService, didArriveAt waypoint: Waypoint) -> Bool {
+        if service.routeProgress.isFinalLeg, service.routeProgress.currentLeg.destination == waypoint {
+            navigationDelegate?.navigationDidArriveAtDestination()
         }
-
-        if progress.currentLegProgress.userHasArrivedAtWaypoint {
-            routeHasChanged = true
-            navigationDelegate?.navigationServiceArrivedAtDestination()
-        }
+        return true
     }
 
     func navigationService(_ service: NavigationService, didRerouteAlong route: MapboxDirections.Route, at location: CLLocation?, proactive: Bool) {
-        routeHasChanged = true
+        navigationDelegate?.navigation(didUpdate: Route(route: route))
     }
 }
 
-
-public protocol NavigationManagerDelegate: class {
-    func navigationService(didUpdate route: MapboxVisionARNative.Route)
-    func navigationServiceArrivedAtDestination()
+public protocol NavigationDelegate: class {
+    func navigation(didUpdate route: MapboxVisionARNative.Route)
+    func navigationDidArriveAtDestination()
 }